refactor: 모달창에 별점 컨테이너 분리
