+++ src/components/Modal/Modal.ts
@@ -1,24 +1,8 @@
 import './style.css';
 import { MovieType } from '../../types/movie';
-import DOM from '../../utils/DOM';
-import { CLOSE_BUTTON, STAR_FILLED, STAR_LINED } from '../../images';
-
-const { $, $$ } = DOM;
-
-export interface MovieScoreEvent extends CustomEvent {
-  detail: {
-    movie: MovieType;
-    score: string;
-  };
-}
-
-const MOIVE_SCORE: Record<string, string> = {
-  '2': '2 최악이예요',
-  '4': '4 별로예요',
-  '6': '6 보통이에요',
-  '8': '8 재미있어요',
-  '10': '10 명작이에요',
-};
+import { CLOSE_BUTTON, STAR_FILLED } from '../../images';
+import ScoreCheckbox from './ScoreCheckbox';
+import { URL } from '../../api/url';
 
 const Modal = {
   render(movie: MovieType) {
@@ -36,7 +20,7 @@ const Modal = {
         </div>
         <div class="modal-main">
             <div class="modal-poster">
-                <img src="https://image.tmdb.org/t/p/w220_and_h330_face/${movie.poster_path}" >
+                <img src="${URL.IMAGE}${movie.poster_path}" >
             </div>
             <div class="modal-details">
                 <div class="movie-info">
@@ -46,30 +30,16 @@ const Modal = {
     )}</p>
                 </div>
                 <p class="movie-text summary">${movie.overview}</p>
-                <div class="score-container">
-                    <h3 class="movie-text">내 별점</h3>
-                    <div class="score-box">
-                        <input type="checkbox" id="star1" name="score" value="2">
-                        <label for="star1"><img src="${STAR_LINED}" alt="별점"></label>
-                        <input type="checkbox" id="star2" name="score" value="4">
-                        <label for="star2"><img src="${STAR_LINED}" alt="별점"></label>
-                        <input type="checkbox" id="star3" name="score" value="6">
-                        <label for="star3"><img src="${STAR_LINED}" alt="별점"></label>
-                        <input type="checkbox" id="star4" name="score" value="8">
-                        <label for="star4"><img src="${STAR_LINED}" alt="별점"></label>
-                        <input type="checkbox" id="star5" name="score" value="10">
-                        <label for="star5"><img src="${STAR_LINED}" alt="별점"></label>
-                    </div>
-                    <h4 class="movie-text">0 점수를 매겨주세요</h4>
-                </div>
+              
             </div>
         </div>
     </div>
     <div class="modal-backdrop"></div>
     `;
 
+    ScoreCheckbox.render(modal, movie);
+
     this.handleModal(modal);
-    this.handleStarCheckbox(modal, movie);
 
     return modal;
   },
@@ -93,41 +63,6 @@ const Modal = {
     window?.addEventListener('keydown', (e) => {
       if (e.key === 'Escape') {
         modal.remove();
-      }
-    });
-  },
-
-  handleStarCheckbox(modal: HTMLDivElement, movie: MovieType) {
-    const checkboxes = $$('.score-box input', modal);
-
-    checkboxes.forEach((checkbox) => {
-      checkbox?.addEventListener('click', (e) => {
-        const element = e.target as HTMLInputElement;
-
-        this.updateMovieScoreUI(element.value);
-
-        const selectMovieScore = new CustomEvent('selectMovieScore', {
-          detail: {
-            movie,
-            score: element.value,
-          },
-        });
-        document.dispatchEvent(selectMovieScore);
-      });
-    });
-  },
-
-  updateMovieScoreUI(score: string) {
-    const starIcons = $$('.score-box img');
-    const scoreMessage = $('.score-container .movie-text:last-child');
-    scoreMessage!.textContent = MOIVE_SCORE[score];
-
-    starIcons.forEach((icon, index) => {
-      const image = icon as HTMLImageElement;
-      if (index + 1 <= Number(score) / 2) {
-        image.src = STAR_FILLED;
-      } else {
-        image.src = STAR_LINED;
       }
     });
   },
