+++ src/components/Modal/style.css
@@ -141,6 +141,10 @@
     width: 60rem;
     height: 40rem;
   }
+
+  .score-container .movie-text {
+    font-size: 1.3rem;
+  }
 }
 
 @media screen and (max-width: 1030px) {
